requirements.txt
alembic==1.7.7
asgiref==3.5.2
astroid==2.11.4
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
blinker==1.4
boto3==1.24.96
botocore==1.27.96
certifi==2020.4.5.1
charset-normalizer==2.1.1
click==8.1.2
colorama==0.4.4
dill==0.3.4
distlib==0.3.5
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-embed-video==1.4.7
django-environ==0.9.0
django-js-asset==2.0.0
django-resized==1.0.2
django-rest-framework==0.1.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.1.1
Flask-Cors==3.0.10
flask-heroku==0.1.9
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==19.5.0
idna==3.4
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.1
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mccabe==0.7.0
openpyxl==3.0.10
packaging==21.3
pep8==1.7.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.5.0
pylint==2.13.8
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.36
SQLAlchemy-serializer==1.4.1
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.4
urllib3==1.26.12
virtualenv==20.16.3
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.0
